+++ changelog.md
@@ -2,6 +2,11 @@
 
 ## Changelog
 
+### 4.5.1, soon...
+
+* fixed: admin notices for missing prerequisites not displaying on non-WooCommerce integrations
+* added: warning notices for prerequisistes of the next major release
+
 ### 4.5.0, 2021-11-14
 
 * changed: update eWAY to Eway and replaced logo images for new Eway branding